+++ src/main/java/ch/admin/bit/jeap/truststoreplugin/mojo/BuildTruststoresMojo.java
@@ -45,6 +45,9 @@ public class BuildTruststoresMojo extends AbstractMojo
     @Parameter(defaultValue = "changeit")
     private String trustStorePassword;
 
+    @Parameter(defaultValue = "CERTIFICATES_REPO_GIT_TOKEN")
+    private String tokenEnvironmentPropertyName;
+
     @Parameter
     private List<String> environments;
 
@@ -94,7 +97,7 @@ public class BuildTruststoresMojo extends AbstractMojo
         try {
             File tempRepoBranchDir = Files.createTempDirectory(branch).toFile();
             FileUtils.forceDeleteOnExit(tempRepoBranchDir);
-            GitCertRepo gitCertRepo = GitCertRepo.of(gitRepoUrl.toURI(), branch);
+            GitCertRepo gitCertRepo = GitCertRepo.of(gitRepoUrl.toURI(), branch, tokenEnvironmentPropertyName);
             getLog().info("Cloning git repo " + gitRepoUrl + " at branch " + branch + " to directory " + tempRepoBranchDir.getAbsolutePath());
             gitCertRepo.cloneTo(tempRepoBranchDir);
             return tempRepoBranchDir;