+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2025-06-20
+
+### Changed
+
+- Enhanced Git clone logic to support private GitHub repositories:
+    - If the environment variable CERTIFICATES_REPO_GIT_TOKEN is set, it is used as a token to authenticate the Git
+      clone operation (e.g., for accessing private GitHub repositories in CI/CD pipelines).
+    - If the token is not provided, the system's installed Git client is used instead, relying on locally configured
+      credentials (e.g., SSH keys or credential helpers) for cloning in developer environments.
+
 ## [1.14.0] - 2025-06-18
 
 ### Changed